refactor(aipanel): 拆出首页视图与对话舞台渲染段
- 新增 components/ai/AIHomeView.tsx：首页渲染段（全局搜索结果/分组 tab 拖拽重排与重命名/多选批量操作条），useMemo 依赖数组原样保留
- 新增 components/ai/AIConversationListRow.tsx：会话列表行（子代理/摘要徽标、临时会话标签、行内操作与 AI_ROW_ACTION_* 常量）
- 新增 components/ai/AIConversationStage.tsx：对话舞台（配色模式横幅/当前对话搜索条/AIChatConversation/协同思考动效）
- AIPanel.tsx 1643 -> 1160 行
