+++ frontend/src/components/AIPanel.tsx
@@ -30,6 +30,8 @@ import { useAIConversationHome } from './ai/useAIConversationHome.ts'
 import { useAIAutoApprovalSettings } from './ai/useAIAutoApprovalSettings.ts'
 import { useAIChatRequests } from './ai/useAIChatRequests.ts'
 import { useAIChatActions } from './ai/useAIChatActions.ts'
+import { renderAIHomeView } from './ai/AIHomeView.tsx'
+import { renderAIConversationStage } from './ai/AIConversationStage.tsx'
 import AIChatConversation from './ai/chat/AIChatConversation.tsx'
 import { getConversationBranchAnchor } from './ai/chat/aiChatMessageTopology.ts'
 import { isCallMyVipProviderHost } from './ai/providerSpecialHosts.ts'
@@ -61,12 +63,6 @@ import { getTemporaryAIConversationSummary, listTemporaryAIConversations as list
 // ============================================================
 // AIPanel 类型契约（props 见 AIPanelProps；内部数据模型见下）
 // ============================================================
-
-const AI_ROW_ACTION_BASE =
-  'w-[26px] h-[26px] inline-flex items-center justify-center rounded-md shadow-none shrink-0 cursor-pointer transition-colors duration-[120ms]';
-const AI_ROW_ACTION_HOVER_ACCENT =
-  'hover:text-accent hover:bg-[rgba(var(--accent-rgb),0.10)] focus-visible:text-accent focus-visible:bg-[rgba(var(--accent-rgb),0.10)]';
-const AI_ROW_ACTION_HOVER_DANGER =
   'hover:text-danger hover:bg-danger-dim focus-visible:text-danger focus-visible:bg-danger-dim';
 
 function AIConversationTabPanel({ width, side, terminalId = 'global', sessionId = '', sessionTerminals = [], workspaceTabId = '', isHomeView = false, isWorkspaceTabActive = true, showComposer = true, initialConversationId = '', tabBar = null, onDevilModeChange, onGoHomeRequested, onOpenConversationRequested, onWorkspaceTabStateChange, addToast }: AIPanelProps) {
@@ -291,370 +287,8 @@ function AIConversationTabPanel({ width, side, terminalId = 'global', sessionId 
 }`
   const configRows = Math.max(configText.split('\n').length, 1)
 
-  const renderedConversationList = useMemo(() => {
-    let content = null
-    const getConversationGroupId = (item: ConversationSummary) => {
-      const ownerId = item.rootConversationId || item.parentConversationId || item.id
-      return conversationOrganizer.assignments[item.id] || conversationOrganizer.assignments[ownerId] || ''
-    }
-    const visibleConversationList = conversationList.filter((item) => {
-      if (conversationFilter === 'archived') return item.archived === true
-      if (item.archived === true) return false
-      if (conversationFilter === 'all') return true
-      const groupId = getConversationGroupId(item)
-      return conversationFilter === 'ungrouped' ? !groupId : groupId === conversationFilter
-    })
-    const displayConversationList = buildAIConversationDisplayList(visibleConversationList)
+  const renderedConversationList = useMemo(() => renderAIHomeView({ t, conversationList, conversationOrganizer, conversationFilter, setConversationFilter, conversationSelectionMode, setConversationSelectionMode, selectedConversationIds, moveToGroupOpen, setMoveToGroupOpen, editingConversationGroupId, editingConversationGroupName, setEditingConversationGroupName, draggingConversationGroupId, dragOverConversationGroupId, setDraggingConversationGroupId, setDragOverConversationGroupId, panelState, globalSearchOpen, globalSearchQuery, setGlobalSearchQuery, normalizedGlobalSearchQuery, globalSearchLoading, globalSearchResults, globalSearchInputRef, conversationGroupRenameInputRef, resetGlobalSearchState, handleOpenGlobalSearch, handleSelectGlobalSearchResult, toggleConversationSelection, clearConversationSelection, handleOpenConversation, handleMakeConversationPermanent, handleOpenConversationFolder, handleRenameConversationTitle, handleDeleteConversation, handleCreateConversationGroup, beginRenameConversationGroup, cancelRenameConversationGroup, commitRenameConversationGroup, reorderConversationGroup, showSystemGroupRenameUnsupported, handleDeleteConversationGroup, handleMoveSelectedConversations, handleSetSelectedArchived, handleDeleteSelectedConversations }), [beginRenameConversationGroup, cancelRenameConversationGroup, clearConversationSelection, commitRenameConversationGroup, conversationFilter, conversationList, conversationOrganizer, conversationSelectionMode, dragOverConversationGroupId, draggingConversationGroupId, editingConversationGroupId, editingConversationGroupName, getLanguage, globalSearchLoading, globalSearchOpen, globalSearchQuery, globalSearchResults, handleCreateConversationGroup, handleDeleteConversation, handleDeleteConversationGroup, handleDeleteSelectedConversations, handleMakeConversationPermanent, handleMoveSelectedConversations, handleOpenConversation, handleOpenConversationFolder, handleOpenGlobalSearch, handleSelectGlobalSearchResult, handleSetSelectedArchived, isDevilMode, moveToGroupOpen, normalizedGlobalSearchQuery, panelState.activeConversationId, reorderConversationGroup, resetGlobalSearchState, selectedConversationIds, showSystemGroupRenameUnsupported, t, toggleConversationSelection])
 
-    if (globalSearchOpen) {
-      content = (
-        <div className="grid min-h-0">
-          <div className="px-3.5 py-2.5 border-b border-line-subtle bg-canvas">
-            <div className="grid grid-cols-[1fr_auto] gap-2 items-center">
-              <input
-                id="ai-panel-main-global-search"
-                name="ai-panel-main-global-search"
-                autoComplete="off"
-                ref={globalSearchInputRef}
-                value={globalSearchQuery}
-                onChange={(event) => setGlobalSearchQuery(event.target.value)}
-                onKeyDown={(event) => {
-                  if (event.key === 'Escape') {
-                    event.preventDefault()
-                    resetGlobalSearchState()
-                  }
-                }}
-                placeholder={t('输入关键词搜索全部对话')}
-                className="h-[34px] w-full rounded-lg border border-line bg-sunken text-primary px-2.5 box-border outline-none"
-              />
-              <button
-                type="button"
-                title={t('关闭搜索')}
-                aria-label={t('关闭搜索')}
-                onClick={resetGlobalSearchState}
-                className="w-[34px] h-[34px] inline-flex items-center justify-center rounded-lg border border-line bg-canvas text-tertiary cursor-pointer"
-              >
-                ×
-              </button>
-            </div>
-          </div>
-          {normalizedGlobalSearchQuery ? (
-            globalSearchLoading ? (
-              <div className="min-h-[calc(100%-101px)] flex items-center justify-center p-5 text-center text-tertiary text-sm leading-[1.8]">
-                {t('加载中...')}
-              </div>
-            ) : globalSearchResults.length === 0 ? (
-              <div className="min-h-[calc(100%-101px)] flex items-center justify-center p-5 text-center text-tertiary text-sm leading-[1.8]">
-                {t('没有找到匹配内容')}
-              </div>
-            ) : (
-              <div className="grid">
-                {globalSearchResults.map((result) => {
-                  const historyTimeParts = buildAIHistoryDisplayTimeParts(result.updatedAt || 0, getLanguage() || 'zh-CN')
-                  const historyRelativeToneStyle = getAIHistoryRelativeTimeToneStyle(result.updatedAt || 0)
-                  return (
-                  <button
-                    key={`${result.conversationId}:${result.messageId}`}
-                    type="button"
-                    onClick={() => {
-                      void handleSelectGlobalSearchResult(result)
-                    }}
-                    className="w-full grid gap-2 py-3 px-3.5 border-0 border-b border-line bg-transparent text-left cursor-pointer"
-                  >
-                    <div className="flex items-center justify-between gap-3">
-                      <div className="min-w-0 text-md font-bold text-primary whitespace-nowrap overflow-hidden text-ellipsis">{result.conversationTitle}</div>
-                      <div className="shrink-0 text-xs text-tertiary">{result.role === 'user' ? t('用户') : t('AI')}</div>
-                    </div>
-                    <div className="text-xs text-muted flex items-center gap-0 flex-wrap">
-                      <span>{historyTimeParts.absoluteText}</span>
-                      {historyTimeParts.relativeText ? (
-                        <span style={historyRelativeToneStyle}>({historyTimeParts.relativeText})</span>
-                      ) : null}
-                    </div>
-                    <div className="text-sm text-secondary leading-[1.6] whitespace-pre-wrap break-words">{result.snippet}</div>
-                  </button>
-                  )
-                })}
-              </div>
-            )
-          ) : (
-            <div className="min-h-[calc(100%-101px)] flex items-center justify-center p-5 text-center text-tertiary text-sm leading-[1.8]">
-              {t('搜索全部对话中的消息')}
-            </div>
-          )}
-        </div>
-      )
-    } else if (displayConversationList.length === 0) {
-      content = (
-        <div className="min-h-[calc(100%-53px)] flex items-center justify-center p-5 text-center text-tertiary text-sm leading-[1.8]">
-          <div className="max-w-[80%] grid gap-0.5">
-            <div>{conversationFilter === 'archived' ? t('当前没有已归档会话') : t('当前分组没有会话')}</div>
-          </div>
-        </div>
-      )
-    } else {
-      content = displayConversationList.map((item) => {
-        const isAgentSubtask = item.relationType === 'agent'
-        const isArchivedAgentSubtask = isAgentSubtask && item.archived === true
-        const isSummarySubtask = item.relationType === 'phase' && item.relationSource === 'summary_condense'
-        const historyTimeParts = buildAIHistoryDisplayTimeParts(item.updatedAt, getLanguage() || 'zh-CN')
-        const historyRelativeToneStyle = getAIHistoryRelativeTimeToneStyle(item.updatedAt)
-        const displayTitle = typeof item.title === 'string'
-          ? item.title.replace(/\s*·\s*摘要子任务\s*$/u, '').replace(/\s*·\s*子代理任务\s*$/u, '').trim()
-          : ''
-        const selected = selectedConversationIds.has(item.id)
-        return (
-          <div
-            key={item.id}
-            className="w-full flex items-center border-b border-line transition-[color,background-color,border-color,opacity,box-shadow] duration-[120ms]"
-            style={{
-              background: selected ? 'rgba(var(--accent-rgb), 0.12)' : panelState.activeConversationId === item.id ? 'rgba(var(--accent-rgb), 0.08)' : 'transparent',
-              borderLeft: panelState.activeConversationId === item.id ? '2px solid var(--accent)' : '2px solid transparent',
-              opacity: item.archived === true ? 0.72 : 1,
-              contentVisibility: 'auto',
-              containIntrinsicSize: '56px',
-              contain: 'layout paint style',
-            }}
-          >
-            {conversationSelectionMode ? (
-              <button
-                type="button"
-                aria-label={selected ? t('取消选择') : t('选择')}
-                aria-pressed={selected}
-                onClick={() => toggleConversationSelection(item.id)}
-                className={cn(
-                  'w-[34px] self-stretch inline-flex items-center justify-center border-0 bg-transparent cursor-pointer shrink-0',
-                  'hover:text-accent hover:bg-[rgba(var(--accent-rgb),0.10)] focus-visible:text-accent focus-visible:bg-[rgba(var(--accent-rgb),0.10)]',
-                  selected ? 'text-accent' : 'text-muted',
-                )}>
-                <span
-                  className={cn(
-                    'w-4 h-4 rounded-sm border inline-flex items-center justify-center text-xs text-white',
-                  )}
-                  style={{ borderColor: selected ? 'var(--accent)' : 'var(--border)', background: selected ? 'var(--accent)' : 'transparent' }}
-                >{selected ? '✓' : ''}</span>
-              </button>
-            ) : null}
-            <button
-              type="button"
-              onClick={() => conversationSelectionMode ? toggleConversationSelection(item.id) : void handleOpenConversation(item.id)}
-              className="flex-1 min-w-0 flex items-center gap-2 px-3 py-2 border-0 bg-transparent text-left cursor-pointer"
-            >
-              <div className="flex-1 min-w-0 grid gap-0.5" style={{ paddingLeft: item.depth > 0 ? `${item.depth * 12}px` : 0 }}>
-                <div className="flex items-center gap-1.5 min-w-0">
-                  {isAgentSubtask ? (
-                    <Tiptop text={t('子代理任务')} placement="top">
-                      <span
-                        aria-label={t('子代理任务')}
-                        className={cn(
-                          'w-[18px] h-[18px] inline-flex items-center justify-center rounded-full shrink-0 border',
-                          isArchivedAgentSubtask
-                            ? 'border-line bg-sunken text-tertiary'
-                            : 'border-[rgba(var(--accent-rgb),0.22)] bg-[rgba(var(--accent-rgb),0.10)] text-accent',
-                        )}
-                      >
-                        <Bot size={11} />
-                      </span>
-                    </Tiptop>
-                  ) : null}
-                  {isSummarySubtask ? (
-                    <Tiptop text={t('摘要子任务')} placement="top">
-                      <span
-                        aria-label={t('摘要子任务')}
-                        className="w-[18px] h-[18px] inline-flex items-center justify-center rounded-full border border-[rgba(var(--accent-rgb),0.22)] bg-[rgba(var(--accent-rgb),0.10)] text-accent shrink-0"
-                      >
-                        <Scissors size={11} />
-                      </span>
-                    </Tiptop>
-                  ) : null}
-                  {item.transient === true ? (
-                    <span title={t('临时会话')} className="shrink-0 px-[5px] py-px rounded-sm bg-[rgba(var(--accent-rgb),0.12)] text-accent text-[10px] font-semibold">
-                      {t('临时会话')}
-                    </span>
-                  ) : null}
-                  <div className={cn(
-                    'min-w-0 text-base leading-tight whitespace-nowrap overflow-hidden text-ellipsis',
-                    isArchivedAgentSubtask ? 'text-secondary' : 'text-primary',
-                    panelState.activeConversationId === item.id ? 'font-semibold' : 'font-medium',
-                  )}>{displayTitle || item.title}</div>
-                </div>
-                <div className="flex items-center gap-0.5 min-w-0 flex-wrap">
-                  <div className="text-xs text-tertiary whitespace-nowrap inline-flex items-center gap-0">
-                    <span>{historyTimeParts.absoluteText}</span>
-                    {historyTimeParts.relativeText ? (
-                      <span style={historyRelativeToneStyle}>({historyTimeParts.relativeText})</span>
-                    ) : null}
-                  </div>
-                  <div className="text-xs text-muted whitespace-nowrap">·{item.messageCount}</div>
-                </div>
-              </div>
-            </button>
-            {!conversationSelectionMode ? <div className="flex items-center gap-1 mr-2.5 shrink-0">
-              {item.transient === true ? (
-                <button
-                  type="button"
-                  title={`${t('临时会话')} → ${t('保存')}`}
-                  aria-label={`${t('临时会话')} → ${t('保存')}`}
-                  onClick={() => void handleMakeConversationPermanent(item.id)}
-                  className={cn(AI_ROW_ACTION_BASE, AI_ROW_ACTION_HOVER_ACCENT, 'text-accent')}
-                >
-                  <ArchiveRestore size={13} />
-                </button>
-              ) : null}
-              {item.transient !== true ? (
-              <button
-                type="button"
-                title={t('打开任务所在文件夹')}
-                aria-label={t('打开任务所在文件夹')}
-                onClick={() => void handleOpenConversationFolder(item.id)}
-                className={cn(AI_ROW_ACTION_BASE, AI_ROW_ACTION_HOVER_ACCENT, 'text-muted')}
-              >
-                <FolderOpen size={13} />
-              </button>
-              ) : null}
-              {item.transient !== true ? (
-              <button
-                type="button"
-                title={t('编辑任务标题')}
-                aria-label={t('编辑任务标题')}
-                onClick={() => void handleRenameConversationTitle(item.id)}
-                className={cn(AI_ROW_ACTION_BASE, AI_ROW_ACTION_HOVER_ACCENT, 'text-muted')}
-              >
-                <Pencil size={13} />
-              </button>
-              ) : null}
-              <button
-                type="button"
-                title={t('删除')}
-                aria-label={t('删除')}
-                onClick={() => {
-                  void handleDeleteConversation(item.id)
-                }}
-                className={cn(AI_ROW_ACTION_BASE, AI_ROW_ACTION_HOVER_DANGER, 'text-muted')}
-              >
-                ×
-              </button>
-            </div> : null}
-          </div>
-        )
-      })
-    }
-
-    return (
-      <div className="flex-1 min-h-0 overflow-y-auto bg-canvas">
-        <div className="px-2.5 py-2 border-b border-line-subtle bg-raised sticky top-0 grid gap-2" style={{ zIndex: Z.STACK }}>
-          <div className="flex items-center justify-between gap-2">
-          <div className="text-sm font-semibold text-secondary">{conversationSelectionMode ? t('已选择 {count} 项').replace('{count}', String(selectedConversationIds.size)) : t('对话历史')}</div>
-          <div className="flex items-center gap-1.5">
-          <button type="button" title={conversationSelectionMode ? t('退出多选') : t('多选')} aria-label={conversationSelectionMode ? t('退出多选') : t('多选')} onClick={() => conversationSelectionMode ? clearConversationSelection() : setConversationSelectionMode(true)} className={cn(
-            'w-7 h-7 inline-flex items-center justify-center rounded-md border cursor-pointer',
-            conversationSelectionMode
-              ? 'border-accent-border bg-[rgba(var(--accent-rgb),0.10)] text-accent'
-              : 'border-line-subtle bg-sunken text-tertiary',
-          )}><CheckSquare size={14} /></button>
-          <button type="button" title={t('新建分组')} aria-label={t('新建分组')} onClick={() => void handleCreateConversationGroup()} className="w-7 h-7 inline-flex items-center justify-center rounded-md border border-line-subtle bg-sunken text-tertiary cursor-pointer"><FolderPlus size={14} /></button>
-          <button
-            type="button"
-            title={t('全局搜索对话')}
-            aria-label={t('全局搜索对话')}
-            onClick={handleOpenGlobalSearch}
-            className={cn(
-              'w-7 h-7 inline-flex items-center justify-center rounded-md border cursor-pointer shrink-0 transition-[color,background-color,border-color,opacity] duration-[80ms]',
-              globalSearchOpen
-                ? 'border-accent-border bg-[rgba(var(--accent-rgb),0.10)] text-accent'
-                : 'border-line-subtle bg-sunken text-tertiary',
-            )}
-          >
-            <Search size={14} />
-          </button>
-          </div>
-          </div>
-          <div role="tablist" aria-label={t('分组')} className="flex gap-1.5 overflow-x-auto [scrollbar-width:none] pb-px">
-            <button role="tab" aria-selected={conversationFilter === 'all'} type="button" onClick={() => { setConversationFilter('all'); clearConversationSelection(); cancelRenameConversationGroup() }} onDoubleClick={showSystemGroupRenameUnsupported} className={cn(
-              'h-[26px] px-[9px] rounded-md border text-xs whitespace-nowrap cursor-pointer shrink-0',
-              conversationFilter === 'all'
-                ? 'border-accent-border bg-[rgba(var(--accent-rgb),0.10)] text-accent'
-                : 'border-line-subtle bg-transparent text-secondary',
-            )}>{t('全部')}</button>
-            {conversationOrganizer.groups.map((group) => {
-              const selected = conversationFilter === group.id
-              const editing = editingConversationGroupId === group.id
-              const dragging = draggingConversationGroupId === group.id
-              const dragOver = dragOverConversationGroupId === group.id && !dragging
-              return editing ? (
-                <input
-                  key={group.id}
-                  ref={conversationGroupRenameInputRef}
-                  aria-label={t('重命名分组')}
-                  value={editingConversationGroupName}
-                  onChange={(event) => setEditingConversationGroupName(event.target.value)}
-                  onBlur={commitRenameConversationGroup}
-                  onKeyDown={(event) => {
-                    if (event.key === 'Enter') { event.preventDefault(); commitRenameConversationGroup() }
-                    if (event.key === 'Escape') { event.preventDefault(); cancelRenameConversationGroup() }
-                  }}
-                  className="h-[26px] px-2 rounded-md border border-accent-border bg-sunken text-primary text-xs outline-2 outline-[rgba(var(--accent-rgb),0.16)] shrink-0"
-                  style={{ width: Math.max(72, Math.min(150, editingConversationGroupName.length * 12 + 24)) }}
-                />
-              ) : (
-                <button
-                  key={group.id}
-                  role="tab"
-                  aria-selected={selected}
-                  draggable
-                  type="button"
-                  onClick={() => { setConversationFilter(group.id); clearConversationSelection() }}
-                  onDoubleClick={() => beginRenameConversationGroup(group.id)}
-                  onContextMenu={(event) => { event.preventDefault(); void handleDeleteConversationGroup(group.id) }}
-                  onDragStart={(event) => { setDraggingConversationGroupId(group.id); event.dataTransfer.effectAllowed = 'move'; event.dataTransfer.setData('text/plain', group.id) }}
-                  onDragOver={(event) => { event.preventDefault(); event.dataTransfer.dropEffect = 'move'; setDragOverConversationGroupId(group.id) }}
-                  onDrop={(event) => { event.preventDefault(); reorderConversationGroup(draggingConversationGroupId || event.dataTransfer.getData('text/plain'), group.id); setDraggingConversationGroupId(''); setDragOverConversationGroupId('') }}
-                  onDragEnd={() => { setDraggingConversationGroupId(''); setDragOverConversationGroupId('') }}
-                  className={cn(
-                    'h-[26px] px-[9px] rounded-md border text-xs whitespace-nowrap shrink-0 transition-[color,background-color,border-color,opacity] duration-[80ms]',
-                    dragOver
-                      ? 'border-accent'
-                      : selected
-                        ? 'border-accent-border'
-                        : 'border-line-subtle',
-                    selected ? 'bg-[rgba(var(--accent-rgb),0.10)] text-accent' : 'bg-transparent text-secondary',
-                  )}
-                  style={{
-                    cursor: dragging ? 'grabbing' : 'grab',
-                    opacity: dragging ? 0.5 : 1,
-                    transform: dragOver ? 'translateX(2px)' : 'none',
-                  }}>
-                  {group.name}
-                </button>
-              )
-            })}
-            <button role="tab" aria-selected={conversationFilter === 'archived'} type="button" onClick={() => { setConversationFilter('archived'); clearConversationSelection(); cancelRenameConversationGroup() }} onDoubleClick={showSystemGroupRenameUnsupported} className={cn(
-              'h-[26px] px-[9px] rounded-md border text-xs whitespace-nowrap cursor-pointer shrink-0',
-              conversationFilter === 'archived'
-                ? 'border-accent-border bg-[rgba(var(--accent-rgb),0.10)] text-accent'
-                : 'border-line-subtle bg-transparent text-secondary',
-            )}>{t('已归档')}</button>
-          </div>
-        </div>
-        {content}
-        {conversationSelectionMode && selectedConversationIds.size > 0 ? (
-          <div className="sticky bottom-0 grid gap-1.5 p-2 border-t border-line bg-raised" style={{ zIndex: Z.STACK + 1 }}>
-            {moveToGroupOpen ? (
-              <div className="flex gap-1.5 overflow-x-auto [scrollbar-width:none]">
-                <Button variant="ghost" size="sm" onClick={() => handleMoveSelectedConversations('')} className="shrink-0">{t('移出分组')}</Button>
-                {conversationOrganizer.groups.map((group) => <Button key={group.id} variant="ghost" size="sm" onClick={() => handleMoveSelectedConversations(group.id)} className="shrink-0">{group.name}</Button>)}
-              </div>
-            ) : null}
-            <div className="flex gap-1.5">
-              <Button variant="ghost" size="sm" onClick={() => setMoveToGroupOpen((current) => !current)} className="flex-1">{t('移动到分组')}</Button>
-              <Button variant="ghost" size="sm" onClick={() => void handleSetSelectedArchived(conversationFilter !== 'archived')} className="flex-1 gap-[5px]">{conversationFilter === 'archived' ? <ArchiveRestore size={13} /> : <Archive size={13} />}{conversationFilter === 'archived' ? t('恢复') : t('归档')}</Button>
-              <Button variant="danger" size="sm" onClick={() => void handleDeleteSelectedConversations()} aria-label={t('删除')} className="w-[34px] p-0"><Trash2 size={13} /></Button>
-            </div>
-          </div>
-        ) : null}
-      </div>
-    )
-  }, [beginRenameConversationGroup, cancelRenameConversationGroup, clearConversationSelection, commitRenameConversationGroup, conversationFilter, conversationList, conversationOrganizer, conversationSelectionMode, dragOverConversationGroupId, draggingConversationGroupId, editingConversationGroupId, editingConversationGroupName, getLanguage, globalSearchLoading, globalSearchOpen, globalSearchQuery, globalSearchResults, handleCreateConversationGroup, handleDeleteConversation, handleDeleteConversationGroup, handleDeleteSelectedConversations, handleMakeConversationPermanent, handleMoveSelectedConversations, handleOpenConversation, handleOpenConversationFolder, handleOpenGlobalSearch, handleSelectGlobalSearchResult, handleSetSelectedArchived, isDevilMode, moveToGroupOpen, normalizedGlobalSearchQuery, panelState.activeConversationId, reorderConversationGroup, resetGlobalSearchState, selectedConversationIds, showSystemGroupRenameUnsupported, t, toggleConversationSelection])
 
   return (
     <AIWorkspaceTabProvider value={{ sessionId: sessionId || '', terminalId: terminalId || '', tabId: workspaceTabId || '' }}>
@@ -733,124 +367,7 @@ function AIConversationTabPanel({ width, side, terminalId = 'global', sessionId 
       />
       {tabBar}
       <div className="flex-1 min-h-0 flex flex-col overflow-hidden">
-        <div data-ai-chat-stage="true" className="flex-1 min-h-0 relative overflow-hidden flex flex-col">
-          {activeConversation || !isHomeView ? (
-            <>
-              {isThemeTuningConversation && !isConversationLoading ? (
-                <div className="px-3 py-2 border-b border-accent-border bg-[rgba(var(--accent-rgb),0.08)] flex items-center justify-between gap-2.5">
-                  <div className="text-sm text-secondary leading-normal">
-                    {t('当前处于配色模式,对话记录不会保存')}
-                  </div>
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    onClick={() => { void handleGoHome() }}
-                    className="shrink-0"
-                  >
-                    {t('退出配色模式')}
-                  </Button>
-                </div>
-              ) : null}
-              {conversationSearchOpen ? (
-                <div className="px-3 py-2 border-b border-line bg-raised grid grid-cols-[1fr_auto_auto_auto_auto] gap-2 items-center">
-                  <input
-                    id="ai-panel-main-conversation-search"
-                    name="ai-panel-main-conversation-search"
-                    autoComplete="off"
-                    ref={conversationSearchInputRef}
-                    value={conversationSearchQuery}
-                    onChange={(event) => setConversationSearchQuery(event.target.value)}
-                    onKeyDown={(event) => {
-                      if (event.key === 'Escape') {
-                        event.preventDefault()
-                        resetConversationSearchState()
-                        return
-                      }
-                      if (event.key === 'Enter') {
-                        event.preventDefault()
-                        handleCycleConversationSearchResult(event.shiftKey ? -1 : 1)
-                      }
-                    }}
-                    placeholder={t('输入关键词搜索当前对话')}
-                    className="h-[34px] w-full rounded-lg border border-line bg-sunken text-primary px-2.5 box-border outline-none"
-                  />
-                  <div className="min-w-12 text-center text-sm text-tertiary tabular-nums">
-                    {conversationSearchResults.length > 0 ? `${conversationSearchIndex + 1}/${conversationSearchResults.length}` : '0/0'}
-                  </div>
-                  <button
-                    type="button"
-                    title={t('上一个搜索结果')}
-                    aria-label={t('上一个搜索结果')}
-                    onClick={() => handleCycleConversationSearchResult(-1)}
-                    disabled={conversationSearchResults.length === 0}
-                    className={cn(
-                      'w-[34px] h-[34px] inline-flex items-center justify-center rounded-lg border border-line bg-canvas',
-                      conversationSearchResults.length > 0 ? 'text-primary cursor-pointer' : 'text-muted cursor-not-allowed',
-                    )}
-                  >
-                    ‹
-                  </button>
-                  <button
-                    type="button"
-                    title={t('下一个搜索结果')}
-                    aria-label={t('下一个搜索结果')}
-                    onClick={() => handleCycleConversationSearchResult(1)}
-                    disabled={conversationSearchResults.length === 0}
-                    className={cn(
-                      'w-[34px] h-[34px] inline-flex items-center justify-center rounded-lg border border-line bg-canvas',
-                      conversationSearchResults.length > 0 ? 'text-primary cursor-pointer' : 'text-muted cursor-not-allowed',
-                    )}
-                  >
-                    ›
-                  </button>
-                  <button
-                    type="button"
-                    title={t('关闭搜索')}
-                    aria-label={t('关闭搜索')}
-                    onClick={resetConversationSearchState}
-                    className="w-[34px] h-[34px] inline-flex items-center justify-center rounded-lg border border-line bg-canvas text-tertiary cursor-pointer"
-                  >
-                    ×
-                  </button>
-                </div>
-              ) : null}
-              <AIChatConversation
-                messages={isConversationLoading ? [] : panelState.messages}
-                sessionId={sessionId}
-                terminalId={terminalId}
-                conversationId={isConversationLoading ? normalizedInitialConversationId || workspaceTabId : activeConversation?.id || workspaceTabId}
-                tabId={workspaceTabId}
-                onSendUserMessage={handleConversationUserMessage}
-                onRetryUserMessage={handleRetryUserMessage}
-                onRetryAssistantMessage={handleRetryAssistantMessage}
-                onEditUserMessage={handleEditUserMessage}
-                onDeleteMessage={handleDeleteMessage}
-                onPreviewRestore={handlePreviewRestore}
-                onPreviewDiffFetch={handlePreviewDiff}
-                onApplyRestore={handleApplyRestore}
-                followupInteractionLocked={collaborationFollowupInteractionLocked}
-                messageActionBarAtBottom={messageActionBarAtBottom}
-                messageNavEnabled={messageNavEnabled}
-                side={side}
-                scrollToBottomSignal={conversationScrollSignal}
-                sendPerfMetricsRef={sendPerfMetricsRef}
-                editingTargetMessageId={composerEditState.mode === 'edit' ? composerEditState.targetMessageId : ''}
-              />
-            </>
-          ) : renderedConversationList}
-          {showAssistantCollaborationActiveImage ? (
-            <video
-              src={assistantThinkingActiveImg}
-              autoPlay
-              loop
-              muted
-              playsInline
-              aria-hidden="true"
-              className="absolute right-[18px] bottom-0 w-[min(32%,180px)] min-w-[120px] max-w-[42vw] max-h-[280px] object-contain pointer-events-none select-none opacity-96 drop-shadow-[0_10px_24px_rgba(0,0,0,0.22)]"
-              style={{ zIndex: Z.STACK }}
-            />
-          ) : null}
-        </div>
+        {renderAIConversationStage({ t, side, sessionId, terminalId, workspaceTabId, isHomeView, activeConversation, isThemeTuningConversation, isConversationLoading, normalizedInitialConversationId, conversationSearchOpen, conversationSearchQuery, setConversationSearchQuery, conversationSearchInputRef, resetConversationSearchState, handleCycleConversationSearchResult, conversationSearchResults, conversationSearchIndex, panelState, handleConversationUserMessage, handleRetryUserMessage, handleRetryAssistantMessage, handleEditUserMessage, handleDeleteMessage, handlePreviewRestore, handlePreviewDiff, handleApplyRestore, collaborationFollowupInteractionLocked, messageActionBarAtBottom, messageNavEnabled, conversationScrollSignal, sendPerfMetricsRef, composerEditState, showAssistantCollaborationActiveImage, renderedConversationList, handleGoHome })}
         {showComposer ? (
         <AIComposer
           onSend={handleComposerSendMessage}
